Binary Semaphore
Resource guarding
SPI or USART even a LED
Not writing in the same or corrupt by interrupts or constext switching
in the middle

It just waits until other things are released.
so wasting systick time.

Cooperative semaphore comes in and saves the yield to other tasks while
waiting status of a semaphore.